move petrinet renderer
