+++ packages/client/hmi-client/src/model-representation/petrinet/petrinet-renderer.ts
@@ -12,8 +12,8 @@ export interface EdgeData {
 	numEdges: number;
 }
 export enum NodeType {
-	State = 'S',
-	Transition = 'T'
+	State = 'state',
+	Transition = 'transition'
 }
 
 const MARKER_VIEWBOX = '-5 -5 10 10';
@@ -86,10 +86,12 @@ export class PetrinetRenderer extends BasicRenderer<NodeData, EdgeData> {
 	}
 
 	renderNodes(selection: D3SelectionINode<NodeData>) {
-		const species = selection.filter((d) => d.data.type === 'S' || d.data.type === NodeType.State);
-		const transitions = selection.filter(
-			(d) => d.data.type === 'T' || d.data.type === NodeType.Transition
-		);
+		const species = selection.filter((d) => d.data.type === NodeType.State);
+		const transitions = selection.filter((d) => d.data.type === NodeType.Transition);
+
+		console.log('\t', species.size());
+		console.log('\t', transitions.size());
+		console.log('\t', selection.data());
 
 		// transitions
 		transitions
@@ -442,11 +444,11 @@ export class PetrinetRenderer extends BasicRenderer<NodeData, EdgeData> {
 
 	getShapeOffset(node: any, angle: number) {
 		switch (node.data.type) {
-			case 'T': {
+			case NodeType.Transition: {
 				// transitions -> squares
 				return { x: node.x, y: node.y };
 			}
-			case 'S': {
+			case NodeType.State: {
 				// species -> circles with multiplier
 				const radius = node.width * 1.5;
 				return { x: node.x + radius * Math.cos(angle), y: node.y + radius * Math.sin(angle) };
@@ -458,7 +460,7 @@ export class PetrinetRenderer extends BasicRenderer<NodeData, EdgeData> {
 
 	addNode(type: string, name: string, pos: { x: number; y: number }) {
 		// FIXME: hardwired sizing
-		const size = type === 'S' ? 60 : 30;
+		const size = type === NodeType.State ? 60 : 30;
 		this.graph.nodes.push({
 			id: `s-${this.graph.nodes.length + 1}`,
 			label: name,
@@ -479,7 +481,7 @@ export class PetrinetRenderer extends BasicRenderer<NodeData, EdgeData> {
 		// FIXME: hardwired sizing
 		const positionX = this.chartSize.width / 2;
 		const positionY = this.chartSize.height / 2;
-		const size = type === 'S' ? 60 : 30;
+		const size = type === NodeType.State ? 60 : 30;
 		this.graph.nodes.push({
 			id: `s-${this.graph.nodes.length + 1}`,
 			label: name,
