+++ packages/client/hmi-client/src/temp/ModelEditor.vue
@@ -74,7 +74,7 @@ import {
 	mathmlToPetri,
 	petriToLatex
 } from '@/petrinet/petrinet-service';
-import { PetrinetRenderer } from '@/petrinet/petrinet-renderer';
+import { PetrinetRenderer } from '@/model-representation/petrinet/petrinet-renderer';
 import Button from 'primevue/button';
 import { useRouter } from 'vue-router';
 import MetadataTable from '@/temp/MetadataTable.vue';
